Create CleanUpDeletedSteamGames.py
*Moved from VariousScripts repo.

I started this because I read somewhere that the Steam Deck leaves behind a mess of files after an uninstallation. I figured since the options online only covered the compatdata and the shader cache, but doesn't address uninstalled game folders left in the common folder, I'd make some scripts of my own and clean everything up. I was surprised to find that Steam has been, at least it seems, cleaning up shader cache and compatdata folders as it goes. So the one script here is to clean up the common folder, which does not *appear* to be done automatically.